Generate overrides section in tempo.yaml only if at least one override is set (#116)
Tempo parses the CLI arguments and sets up the config object with values
from the CLI arguments or their default values. Later, this config
object is overlayed with the unmarshalled yaml configuration file.

An empty overrides key in the yaml:
```
overrides:
```
sets overrides to nil instead of an empty map. `yaml.Unmarshal`
unmarshals nil values to structs by setting every struct member
to the zero value.

This results in overwriting the default values, for example the setting
`ingestion_rate_limit_bytes` will be set to 0 instead of the default
value as provided by tempo.