Restart pods when the tempo configuration changes
This change adds the checksum of the tempo configuration file to the
PodTemplateSpec of each Deployment and StatefulSet. If the tempo
configuration changes, the checksum will be updated, and the Deployments
and StatefulSet will trigger a recreation of the pods. This is required
for tempo to load the new configuration.

Resolves: #137
Signed-off-by: Andreas Gerstmayr <agerstmayr@redhat.com>