Validate storage secret in validation webhook if it exists
This partially reverts https://github.com/os-observability/tempo-operator/pull/167
and brings back validation of the storage secret in the validation
webhook.

However, a missing secret does not fail the validation. Instead, the
cluster will be in a degraded state until the storage secret is set.
This is helpful in case manifests get applied out of order, i.e. the
Tempo CR gets created before the storage secret.

Signed-off-by: Andreas Gerstmayr <agerstmayr@redhat.com>