Validate storage secret in webhook if it exists (#192)
* Validate storage secret in validation webhook if it exists

This partially reverts https://github.com/os-observability/tempo-operator/pull/167
and brings back validation of the storage secret in the validation
webhook.

However, a missing secret does not fail the validation. Instead, the
cluster will be in a degraded state until the storage secret is set.
This is helpful in case manifests get applied out of order, i.e. the
Tempo CR gets created before the storage secret.

Signed-off-by: Andreas Gerstmayr <agerstmayr@redhat.com>

* Fix degraded to degraded status condition transition

Previously the reason wasn't updated if the condition was already
degraded. Added a new unit test to check that case.

Signed-off-by: Andreas Gerstmayr <agerstmayr@redhat.com>

---------

Signed-off-by: Andreas Gerstmayr <agerstmayr@redhat.com>