Always set all status condition values in the tempostack_status_condition metric
In some cases not all status condition metrics were updated, because for example:
A TempoStack CR gets created with an invalid storage secret (creating a Degraded status condition).
Later this CR is deleted, a storage secret is created and a new TempoStack instance is created.
Then this TempoStack instance doesn't have the degraded condition in the status.Conditions list.

Signed-off-by: Andreas Gerstmayr <agerstmayr@redhat.com>