Support upgrading previously unmanaged instances
Previously, the operator performed upgrades only at operator startup.
This is a problem if a TempoStack instance gets switched from Unmanaged
to Managed - in this case, the operator should perform the upgrade once
the TempoStack is managed again.

This commit runs the upgrade process in the reconcile loop to cater to
the above use case.

Resolves: #478
Signed-off-by: Andreas Gerstmayr <agerstmayr@redhat.com>